+++ sql-planner/src/ir/distribution/tests.rs
@@ -69,7 +69,7 @@ fn projection_any_dist_for_expr() {
     insta::assert_snapshot!(plan.as_explain().unwrap(), @r#"
     projection (sum(("count_1"::int))::int -> "col_1")
         motion [policy: full, program: ReshardIfNeeded]
-            projection (count(("test_space"."id"::int))::int -> "count_1")
+            projection (count(("test_space"."id"::int::int))::int -> "count_1")
                 scan "test_space"
     execution options:
         sql_vdbe_opcode_max = 45000
