+++ sql-planner/src/ir/explain/tests/query_explain.rs
@@ -196,7 +196,7 @@ fn test_query_explain_11() {
     projection ("gr_expr_1"::string -> "a", sum(("count_1"::int))::int -> "col_1")
         group by ("gr_expr_1"::string) output: ("gr_expr_1"::string -> "gr_expr_1", "count_1"::int -> "count_1")
             motion [policy: full, program: ReshardIfNeeded]
-                projection ("unnamed_subquery_1"."a"::string -> "gr_expr_1", count(("unnamed_subquery_1"."b"::int))::int -> "count_1")
+                projection ("unnamed_subquery_1"."a"::string -> "gr_expr_1", count(("unnamed_subquery_1"."b"::int::int))::int -> "count_1")
                     group by ("unnamed_subquery_1"."a"::string) output: ("unnamed_subquery"."e"::int -> "e", "unnamed_subquery"."f"::int -> "f", "unnamed_subquery_1"."a"::string -> "a", "unnamed_subquery_1"."b"::int -> "b")
                         join on "unnamed_subquery"."e"::int = "unnamed_subquery_1"."b"::int
                             scan "unnamed_subquery"
